requirements.txt
click
numpy
matplotlib
biopython
statsmodels
sklearn
pysam
pyfaidx